Problem with material pdf is not shown #1674

+++ webapp/src/app/pages/components/materials/material/material.component.ts
@@ -99,7 +99,7 @@ export class MaterialComponent implements OnInit, OnDestroy, AfterViewChecked {
   @Output() conceptMapEvent: EventEmitter<boolean> = new EventEmitter();
   @Output() selectedToolEvent: EventEmitter<string> = new EventEmitter();
   cmSelected = false;
-
+  showPdfErrorMessage = false;
   showDialog: boolean = false;
 
   showModeratorPrivileges: boolean;
@@ -227,6 +227,7 @@ export class MaterialComponent implements OnInit, OnDestroy, AfterViewChecked {
               if (this.selectedChannel._id === this.channels['_id']) {
                 this.materials = this.channels['materials'];
                 this.materials.forEach((material) => {
+                  
                   this.showFullMap[material._id] = false;
                 });
               } else {
@@ -285,6 +286,9 @@ export class MaterialComponent implements OnInit, OnDestroy, AfterViewChecked {
     this.lastTimeCourseMapperOpened$ = this.store.select(
       getLastTimeCourseMapperOpened
     );
+    this.pdfViewService.pdfError$.subscribe((error) => {
+      this.showPdfErrorMessage = error;
+    });
   }
   toggleFullMaterialName(materialId: string, event: MouseEvent): void {
     // event.preventDefault();
@@ -304,7 +308,9 @@ export class MaterialComponent implements OnInit, OnDestroy, AfterViewChecked {
     if (words.length <= limit) return text;
     return words.slice(0, limit).join(' ') + '...';
   }
-
+  dismissError() {
+    this.pdfViewService.clearError();
+  }
   getMaterialActivityIndicator(materialId: string) {
     return combineLatest([
       this.allNotifications$,
@@ -478,6 +484,7 @@ export class MaterialComponent implements OnInit, OnDestroy, AfterViewChecked {
     // }
     if (!this.selectedMaterial) return;
     switch (this.selectedMaterial.type) {
+      
       case 'pdf':
         this.pdfViewService.setPageNumber(1);
         let url =
@@ -565,7 +572,7 @@ export class MaterialComponent implements OnInit, OnDestroy, AfterViewChecked {
         },
       });
     } else if (
-      (this.selectedMaterial.type == 'pdf' && this.selectedMaterial.url) ||
+      (this.selectedMaterial.type == 'pdf' && this.selectedMaterial.url && !this.showPdfErrorMessage) ||
       (this.selectedMaterial.type == 'video' && this.selectedMaterial.url == '')
     ) {
       this.materialService.deleteFile(this.selectedMaterial).subscribe({
@@ -599,6 +606,7 @@ export class MaterialComponent implements OnInit, OnDestroy, AfterViewChecked {
         ]);
 
         // e.index = 1
+        this.showPdfErrorMessage=false
         this.showInfo('Material successfully deleted!');
       },
       error: (err) => {
